backend/prods_fastapi/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pandas==2.1.3
numpy==1.25.2
opencv-python==4.8.1.78
Pillow>=10.0.0,<10.1.0
scikit-learn==1.3.2
scipy==1.11.4
pydantic>=2.7.0,<3.0.0
pydantic-settings==2.8.1
python-multipart==0.0.6
webcolors==1.13

# New dependencies for performance optimization
redis==4.6.0
celery==5.3.4
asyncpg==0.29.0
aiofiles==23.2.1
sqlalchemy[asyncio]==2.0.23
asyncio-pool==0.6.0
cachetools==5.3.2
aiocache==0.12.2
motor==3.3.2

# Background task processing
dramatiq==1.14.2
dramatiq[redis]==1.14.2

# Load balancing and monitoring
prometheus-client==0.19.0
py-healthcheck==1.10.1
psutil==7.0.0
sentry-sdk[fastapi]==1.38.0

# Image processing optimization
opencv-python-headless==4.8.1.78
imageio==2.31.6
mediapipe==0.10.7
# Face detection alternatives - using MediaPipe (already used) and OpenCV
# No face-recognition to avoid dlib compilation issues

# CDN and file handling
cloudinary==1.40.0
boto3==1.34.0
minio==7.2.0

# Rate limiting
slowapi==0.1.9
